# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddWorkerToShopRequest:
    properties:
      coffee_shop_id:
        type: string
      worker_id:
        type: string
    required:
    - coffee_shop_id
    - worker_id
    type: object
...
  dto.CoffeeShopResponse:
    properties:
      address:
        type: string
      contacts:
        type: string
      id:
        type: string
      name:
        type: string
      rules:
        type: string
      welcome_message:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  dto.WorkerCoffeeShopResponse:
    properties:
      coffee_shop:
        $ref: '#/definitions/dto.CoffeeShopResponse'
      id:
        type: string
      worker:
        $ref: '#/definitions/dto.UserResponse'
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /admin/coffee-shops/{id}/workers:
    get:
      description: Retrieves a paginated list of users working in a specific coffee
        shop. Requires admin access to the coffee shop.
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List workers in a coffee shop
      tags:
      - worker-coffee-shops
...
  /admin/worker-coffee-shops:
    post:
      consumes:
      - application/json
      description: Adds a user as a worker to a specific coffee shop. Requires creator
        or admin access to the coffee shop.
      parameters:
      - description: Worker and Coffee Shop IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddWorkerToShopRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.WorkerCoffeeShopResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a worker to a coffee shop
      tags:
      - worker-coffee-shops
  /admin/worker-coffee-shops/{id}:
    delete:
      description: Removes a worker-coffee shop relationship by its ID. Requires creator
        or admin access to the coffee shop.
      parameters:
      - description: Worker Coffee Shop Relationship ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove a worker from a coffee shop
      tags:
      - worker-coffee-shops
...
  /users/{id}/coffee-shops:
    get:
      description: Retrieves a paginated list of coffee shops a specific user works
        for. A user can only view their own list.
      parameters:
      - description: Worker User ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CoffeeShopResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List coffee shops for a worker
      tags:
      - worker-coffee-shops
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"